ci: refactor GitHub Actions workflow file (#1249)
* Clean up the workflow file

This removes the trailing whitespace from the workflow, removes an
unused step and fixes up some yaml formatting.

* Make the caching more aggressive

This makes the caching more aggressive by allowing for partial
restoration of the cached data and use the cache for more jobs.

* Run all the tests in parallel

Have stopped the testsuites from requiring the compile checks, Github
Actions doesn't mind that we are running multiple tests in parallel and
this does over a bit of a speed up.

* Separate Security Audit into a separate workflow

Have moved the Cargo Audit check into it's own workflow and switched it
to use the `actions-rs/audit-check` to simplify the setup. Have also
added in a daily security scan.